# backend/fire_spread/openapi.yaml
openapi: 3.1.0
info:
  title: Fire Spread API
  description: >
    Self-hosted ELMFIRE fire-spread service for Catalonia. Given a point ignition,
    runs a Monte Carlo ensemble of ELMFIRE simulations on open Catalan data (50 m
    cells, 40x40 km domain placed downwind of the ignition) and returns a lat/lon grid with minute-level arrival
    times and per-cell uncertainty (burn probability, P10/P90 arrival).

    Mounted at /fire inside the FireProtector API (backend/app), whose global
    error handlers apply: every error body is {"error": {"code", "message"}}
    rather than {"detail"} and 422 is reported as 400 (see backend/README.md,
    "Errors"). The status codes below are what the router itself raises.
  version: "2.0.0"

servers:
  - url: /fire

tags:
  - name: fire-spread

paths:
  /arrival-grid:
    get:
      tags: [fire-spread]
      summary: Simulate a point ignition and return the arrival grid
      description: >
        Synchronous: the response is the prediction (typically 10-60 s). One request
        = one ELMFIRE ensemble run. Concurrency is limited; a 503 means retry later.
        The grid is cropped to the burned extent (plus a small margin) so its size
        depends on how far the fire spreads.
      operationId: arrivalGrid
      parameters:
        - name: lat
          in: query
          required: true
          description: Ignition latitude (WGS84).
          schema: { type: number, format: double, minimum: -90, maximum: 90 }
        - name: lon
          in: query
          required: true
          description: Ignition longitude (WGS84).
          schema: { type: number, format: double, minimum: -180, maximum: 180 }
        - name: durationHours
          in: query
          required: false
          description: Simulation horizon in hours.
          schema: { type: integer, minimum: 1, maximum: 48, default: 24 }
        - name: ensembleMembers
          in: query
          required: false
          description: >
            Number of ELMFIRE cases. 1 = deterministic (burnProbability is 0/1 and
            P10 = median = P90). Each case runs on its own NWP ensemble member (wind,
            temperature, humidity, rain) plus small perturbations of 1-h fuel moisture
            and the spread-rate adjustment; without an ensemble the deterministic wind
            is perturbed with the ensemble spread instead.
          schema: { type: integer, minimum: 1, maximum: 64, default: 16 }
        - name: startTime
          in: query
          required: false
          description: >
            Ignition time (ISO 8601). Weather band 1 is the hour containing it and the
            minute offset becomes the simulation start. Default now.
          schema: { type: string, format: date-time }
        - name: seed
          in: query
          required: false
          description: ELMFIRE random seed. Default derived from lat/lon/startTime (reproducible).
          schema: { type: integer, minimum: 1 }
        - name: spotting
          in: query
          required: false
          description: Enable ember transport (spot fires). Slower; default from server settings (off).
          schema: { type: boolean }
        - name: debug
          in: query
          required: false
          description: Include run directory, timings and ELMFIRE stdout tail.
          schema: { type: boolean, default: false }
      responses:
        "200":
          description: Arrival grid computed successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArrivalGrid"
        "422":
          description: >
            Validation error, or the ignition is outside the static-data coverage
            (Catalonia) or on a non-burnable cell (urban, water, agriculture, barren).
            Reported as 400 by the FireProtector API.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/HTTPError" }
        "500":
          description: ELMFIRE failed (see detail / run directory).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/HTTPError" }
        "502":
          description: Weather provider (Open-Meteo) failure.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/HTTPError" }
        "503":
          description: >
            All simulation slots busy (retry later), or the static tier has not been
            built yet (scripts/setup_fire_data.sh).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/HTTPError" }
        "504":
          description: ELMFIRE exceeded its time limit.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/HTTPError" }

  /health:
    get:
      tags: [fire-spread]
      summary: Liveness + whether the ELMFIRE binary and static data are present
      operationId: health
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: { type: string }
                  elmfire: { type: boolean }
                  dataDir: { type: string }
                  dataPresent: { type: boolean }

  /data-info:
    get:
      tags: [fire-spread]
      summary: Manifest of the static data tier (grid, sources, dates, licenses)
      operationId: dataInfo
      responses:
        "200":
          description: Contents of data/fire_spread/catalonia/manifest.json
          content:
            application/json:
              schema: { type: object }
        "404":
          description: Static data not prepared.

components:
  schemas:
    NullableIntGrid:
      type: array
      description: Rows south-to-north, each a list of columns west-to-east.
      items:
        type: array
        items:
          type: [integer, "null"]
          minimum: 0
    NullableNumberGrid:
      type: array
      description: Rows south-to-north, each a list of columns west-to-east.
      items:
        type: array
        items:
          type: [number, "null"]
          minimum: 0

    ArrivalGrid:
      type: object
      description: >
        A regular lat/lon grid anchored at its south-west corner. Cell (row, col)
        covers [originLon + col*cellDegLon, +cellDegLon) x [originLat + row*cellDegLat,
        +cellDegLat). The ignition point sits at a cell centre. null = no ensemble
        member reached the cell within the horizon, or the cell is outside data coverage.
      required:
        - originLat
        - originLon
        - cellDegLat
        - cellDegLon
        - cellSizeM
        - durationMinutes
        - ensembleMembers
        - arrivalHours
        - arrivalMinutes
        - arrivalMinutesP10
        - arrivalMinutesP90
        - burnProbability
        - weather
      properties:
        originLat:
          type: number
          format: double
          description: Latitude of the south-west corner of cell [0][0].
        originLon:
          type: number
          format: double
          description: Longitude of the south-west corner of cell [0][0].
        cellDegLat:
          type: number
          format: double
          description: Cell height in degrees latitude (~cellSizeM).
        cellDegLon:
          type: number
          format: double
          description: Cell width in degrees longitude (varies with ignition latitude).
        cellSizeM:
          type: number
          description: Nominal cell size in metres (the static tier's resolution, 50 or 30).
        durationMinutes:
          type: integer
          description: Simulation horizon in minutes.
        ensembleMembers:
          type: integer
          description: Number of ELMFIRE cases actually aggregated.
        arrivalHours:
          allOf: [{ $ref: "#/components/schemas/NullableIntGrid" }]
          description: >
            ceil(arrivalMinutes / 60): first hour (0 = ignition cell, 1..N) the
            ensemble-median fire reaches the cell. Kept for backward compatibility.
        arrivalMinutes:
          allOf: [{ $ref: "#/components/schemas/NullableNumberGrid" }]
          description: Ensemble median arrival time in minutes (over members that burned the cell).
        arrivalMinutesP10:
          allOf: [{ $ref: "#/components/schemas/NullableNumberGrid" }]
          description: 10th percentile arrival time (earliest plausible).
        arrivalMinutesP90:
          allOf: [{ $ref: "#/components/schemas/NullableNumberGrid" }]
          description: 90th percentile arrival time (latest plausible).
        burnProbability:
          type: array
          description: Fraction of ensemble members in which the cell burned (0-1).
          items:
            type: array
            items:
              type: [number, "null"]
              minimum: 0
              maximum: 1
        weather:
          $ref: "#/components/schemas/WeatherSummary"
        physics:
          type: object
          description: >
            Model switches that shaped this run: spotting, barriers (OSM fire breaks
            available), diurnalAdjustment, sunriseUtc/sunsetUtc (h), weatherGrid (points
            per side), cellSizeM, ignitionOffsetS (seconds into the ignition hour).
          additionalProperties: true
        zone:
          $ref: "#/components/schemas/ZoneInfo"
        debug:
          $ref: "#/components/schemas/DebugInfo"

    WeatherSummary:
      type: object
      required: [source, windSpeedAvgMs, windDirectionAvg]
      properties:
        source: { type: string, description: "e.g. open-meteo:best_match+icon_eu_eps" }
        windSpeedAvgMs: { type: number, description: Mean 10 m wind speed over the horizon (m/s). }
        windDirectionAvg: { type: number, description: Circular mean wind direction (deg, from). }
        windSpeedSigmaMs: { type: [number, "null"], description: Ensemble wind-speed spread used for perturbations. }
        windDirectionSigmaDeg: { type: [number, "null"] }
        fuelMoisture1hAvgPct: { type: [number, "null"], description: Horizon mean of the time-lagged 1-h dead fuel moisture at the ignition. }
        fuelMoisture100hAvgPct: { type: [number, "null"] }
        liveHerbaceousPct: { type: [number, "null"], description: Live herbaceous moisture used (monthly climatology unless configured). }
        liveWoodyPct: { type: [number, "null"] }
        foliarMoisturePct: { type: [number, "null"], description: Canopy foliar moisture (crown-fire initiation). }
        weatherMembers: { type: [integer, "null"], description: NWP ensemble members driving the cases (1 = perturbed deterministic). }

    ZoneInfo:
      type: object
      description: Bombers/Interior ZHR 2014 fire-regime zone containing the ignition (metadata only).
      properties:
        name: { type: [string, "null"] }
        dominantFireType: { type: [string, "null"] }
        designFires:
          type: object
          additionalProperties: { type: number }
        properties:
          type: object
          description: Raw attributes of the zone feature.

    DebugInfo:
      type: object
      properties:
        runId: { type: string }
        runDir: { type: string }
        timings:
          type: object
          additionalProperties: { type: number }
        elmfireStdoutTail: { type: string }

    HTTPError:
      type: object
      required: [detail]
      properties:
        detail:
          description: >
            Human-readable error message (or FastAPI validation detail list). Inside the
            FireProtector API the same text arrives as error.message.